+++ ChangeLog.md
@@ -1,6 +1,7 @@
 
 # Current
 
+ * added jacobian for temperature-dependent conductivity
  * updated LE10 benchmark
  * read time-dependent functions in `.msh` files
  * detect/handle hanging nodes